minor cleanup
